mcp-orchestrator/requirements.txt
# Core dependencies
pydantic>=2.0.0
asyncio-mqtt>=0.16.0
aiohttp>=3.9.0
websockets>=12.0

# State management
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Workflow integration
httpx>=0.26.0

# Voice processing
openai>=1.0.0  # For Whisper API
google-cloud-texttospeech>=2.14.0

# Logging and monitoring
structlog>=24.1.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
